requirements.txt
datasets
sacrebleu 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
bitsandbytes==0.40.2
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
trl
evaluate==0.4.0
peft
pymongo==4.2.0
python-dotenv==0.21.0
pymongo[srv]
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.40.2
transformers==4.31.0
trl==0.4.7
-e .